Small upd

+++ nuTerra/shaders/Terrain_shaders/Terrain.frag
@@ -92,7 +92,7 @@ in VS_OUT {
     vec2 UV;
     vec2 Global_UV;
     float ln;
-    flat float map_id;
+    flat uint map_id;
 } fs_in;
 
 /*===========================================================*/
