+++ nuTerra/shaders/mapMixing/t_mixer.frag
@@ -84,7 +84,7 @@ in VS_OUT {
     vec3 worldPosition;
     vec2 UV;
     vec2 Global_UV;
-    flat float map_id;
+    flat uint map_id;
 } fs_in;
 
 /*===========================================================*/
